requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0

# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Configuration
pyyaml==6.0.1
